proc: make switch_to(NULL) a noop
Previously, it would switch out of any address space and into the
boot_pgdir.  However, it wouldn't permanently leave the process's
address space, and boot_pgdir is mapped in all address spaces.  No one
was using it in this manner.

Although I can think of reasons to do this, no one was doing it, and
it's more convenient to have NULL be a noop.  In particular, this is for
PCI code that may or may not need to switch into a process's address
space.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>