pci: add a helper to lookup a pci_device by string
This is for #device code that often wants to find a pci_device that
userspace specified by string.  The string is of the form 00:00.0.  You
can skip leading zeros.

This one throws on error, since the intent is that user-interface code
will use it.  The older match_tbdf() is used in a few internal places,
with a custom format for the int argument (MKBUS()).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>