dma: add user DMA arenas
These are arenas of user virtual addresses.  The kernel can allocate
them.  When we run a kernel driver for a device in the user's address
space, these arenas will be used for DMA addresses.  i.e. addresses the
device can access.

I might be able to use these for UCQs or something in the future.  In
essence, they are kernel-allocated and kernel-freed userspace memory.

Arguably, we could try to use arenas to do the user VA allocation, and
then hook in to the specific item (e.g. anon memory, file cache,
preexisting page) in a second step.  In that sense, do_mmap() would call
the UVA arena allocator.

For now, just have the arena be a wrapper interface around do_mmap() is
fine. (arena->do_mmap, versus do_mmap->arena + other magic).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>