proc: iommu: flush the iotlb during shootdowns
IOMMUs are like remote cores.  When we need a generic TLB shootdown, we
also need to shootdown the IOTLB for any IOMMUs where the process's
address space is loaded.

If you don't use device assignment, this is a noop.  If you do, it's
pretty expensive: 800 ns or so.

Part of the reason it is worse than a regular IPI shootdown is that we
have to wait for a response from the IOMMU hardware.  With IPIs, we know
that the core was using the page table (e.g. userspace or a syscall that
wasn't in IRQ context) immediately enters the handler and stops using
the page table.  In short, we stop the processor from processing and
using the table.  Then the actual flush happens.  With the IOMMU, we
can't stop the device from processing, so we have to wait until the
flush completes.

Just like IPIs, these are relatively expensive, and any tricks we want
to do to amortize or batch up TLB shootdowns will apply equally to the
IOTLB.  e.g. "defer reuse / freeing of pages and PTEs".

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>