ioat: support running in a process's address space
Whenever a device is assigned to a process, it's DMA *allocations* come
from the process's dma_arena.  These addresses are user addresses: both
driver and device.  So when we operate in the driver and touch those
bits, we need to be in the user's address space.

Note that dma_map_page() will not return a user-address.  We could try
to wire that up, but it's a bit of a pull.  The issue is that fresh
allocations are easy, but taking an existing kernel page and mapping it
into the user is a little tougher.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>