cbdma: overhauled #cbdma
This uses the dma engine API, and the Linux IOAT driver beneath that.

For whatever reason, we auto-assign 0:4.3 to the kernel, and that is the
only device used by #cbdma/ktest.

	cat cbdma/ktest 		# tests 0:4.3 in the kernel
	echo 00:04.3 0 > iommu/detach 	# detaches from the kernel
        echo 00:04.4 PID > iommu/detach # attaches to PID
        ucbdma 0:4.3 			# attach, test, detach
        ucbdma 0:4.4

ucbdma spits out all of the iommu mappings when it runs.  It's basically
the ktest "do a single memcpy and don't break anything" type of test.
i.e. it doesn't test stuff like "can i do a ucbdma without assigning" or
whatnot.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>