iommu: rewrite device assignment
This was a mess.  The data structures and connections between processes,
pci_devices, and iommus were all a mess.  The pci device assignment
wasn't integrated with the iommu, etc.

The main thing is that processes now have lists of their iommus, which
will make things less painful when we do an IOTLB flush.  It's still
painful, but not as bad.

All of the assignment stuff was changed, so now you call the pci
functions, e.g. pci_assign_device(), which will internally deal with the
iommu.

When processes are destroyed, we tear down any assigned device.  In my
current code, all in-kernel device users hold the pdev qlock.  For
example, when #cbdma tries to use the IOAT driver, it qlocks the device.
I think we'd be OK if the device was unassigned out from under the
driver, in that the IOMMU would protect us from any wild writes, but the
kernel driver would likely get wedged.

I have a bunch of changes to #cbdma, but the easiest thing right now is
to just not build it briefly.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>