tests/lock_test: add a linux kernel module for MCS locks
This uses the in-kernel "queue" locks, which are basically special MCS
locks.  The test runs in the kernel, and we extract the information to
userspace for processing, using the same commands and analysis as for
user-space spinlocks.

Right now, there's only the one kernel lock.  We could add more, like
regular spinlocks and whatnot.  That way, you'd be able to test a
variety of spinlocks, with the low-interference you'd expect from
running inside the kernel (irqs disabled, no resched while holding
locks, etc).

One reason to add other locks would be to test *our own* locks, instead
of Linux's locks.  In that sense, the module would be a platform for
benchmarking lock performance, generally.  Versus just testing Linux
implementations.   Especially since, from early numbers, our Linux
userspace MCS locks (which are basically the ones we coded for Akaros)
are better than the in-kernel MCS locks (queue_lock()).

To build this, you'll need to set up your LINUX_KDIR variable, but also
hack the source to export a symbol.  Oh well.  You'll also need to be
running that kernel.  You should be able to hack this up for a running
machine.  (e.g. cat the symtab, find the function, cast that arbitrary
number to a function pointer, and cross your fingers).

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>