Fix Kbuild's rebuilding issue
Kbuild was rebuilding the entire kernel on each make.  This might be due
to a recent change in GNU make.  When running with make V=2, the change
was due to a command line difference.  The difference (via hacking
Kbuild.include) was KBUILD_STR, where the \#s was getting commented out
from one of the versions of the cmd_ arguments.

Anyway, the simplest thing was to remove KBUILD_STR, which Linux already
did.  This commit basically "side-ports" Linux commit b42841b7bb62
("kbuild: Get rid of KBUILD_STR"), plus adding 'quote'.

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>