x86: vmm: support dst_shorthand APIC IPIs
There are a couple bits in the ICR that specify things like 'broadcast'
and 'self'.  Linux uses these if nr_cpus > 2.

Note that we have to do the check in the kernel too.  The kernel gets
first dibs on resolving an vmexit.  That took a couple hours to figure
out...

Signed-off-by: Barret Rhoden <brho@cs.berkeley.edu>